Move regsitry and cosign utilities to `dangerzone/updater/*`.
Placing these inside the `dangerzone` python package enables an
inclusion with the software itself, and also makes it possible for
end-users to attest the image.
