+++ dangerzone/updater/utils.py
@@ -1,0 +1,3 @@
+def write(file, content: bytes | str):
+    file.write(content)
+    file.flush()
